indexer/requirements.txt
chromadb==0.5.7
sentence-transformers==2.7.0
pymupdf==1.24.9
pydantic==2.7.4
tqdm==4.66.4
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pandas==2.2.2
extract-msg==0.48.6
beautifulsoup4==4.12.3
lxml==5.2.2
olefile==0.47
